+++ lib/sixarm_ruby_markdown.rb
@@ -5,7 +5,6 @@ Please see README
 
 require "ostruct"
 require "sixarm_ruby_equal_instance_variables"
-require "sixarm_ruby_file_rewrite"
 
 require_relative "sixarm_ruby_markdown/sixarm/markdown/headline"
 require_relative "sixarm_ruby_markdown/sixarm/markdown/file"