+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Changes
 
+* 2018-12-23 3.1.0 Add trim of atypical characters
+* 2018-09-23 3.0.2 Bump gems
+* 2018-09-23 3.0.1 Fix gem dependency
 * 2018-06-22 3.0.0 Update to Ruby 2.5, testing gems, etc.
 * 2018-02-15 2.1.3 Fix headline comparison via equal method
 * 2018-02-14 2.1.1 Fix headline parsing of newline